# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EVS Meter Tools API
  version: 1.0.0
  description: API documentation for EVS top-up WebApp, website lookup, and support routes.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /website/lookup:
    get:
      tags: [Website]
      summary: Look up meter balance, usage, or recent top-ups
      parameters:
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [balance, usage, topups]
            default: balance
        - $ref: "#/components/parameters/MeterId"
        - name: hostel
          in: query
          required: false
          schema:
            type: string
            enum: [cp2, cp2nus, sutd]
          description: Optional system hint. SUTD usage history is not currently available.
      responses:
        "200":
          description: Lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsiteLookupResponse"
        "400":
          description: Invalid lookup mode, hostel, or meter ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
        "502":
          description: Upstream lookup failed
  /website/feedback:
    post:
      tags: [Website]
      summary: Submit website feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                message:
                  type: string
                  maxLength: 2000
                contact:
                  type: string
                  maxLength: 160
      responses:
        "200":
          description: Feedback accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid rating
        "429":
          description: Rate limited
...
components:
  parameters:
    MeterId:
      name: txtMtrId
      in: query
      required: true
      schema:
        type: string
        pattern: "^\\d{8}$"
        example: "12345678"
      description: Eight-digit EVS meter ID.
...
  schemas:
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        code:
          type: string
          example: INVALID_AMOUNT
        error:
          type: string
...
    WebsiteLookupResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        mode:
          type: string
          enum: [balance, usage, topups]
        meterId:
          type: string
        hostel:
          type: string
        address:
          type: string
        balance:
          type: string
        checkedAt:
          type: string
          format: date-time
        usage:
          type: object
          description: Present when mode=usage.
        topups:
          type: object
          description: Present when mode=topups.